Move validators and errors to related folders
